fix file not found by renaming GettingStarted file and redirecting route in Reading.jsx
